#813 企业微信中部门id类型改为Long，以容纳更大的数值

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/bean/WxCpDepart.java
@@ -1,9 +1,9 @@
 package me.chanjar.weixin.cp.bean;
 
+import java.io.Serializable;
+
 import lombok.Data;
 import me.chanjar.weixin.cp.util.json.WxCpGsonBuilder;
-
-import java.io.Serializable;
 
 /**
  * 微信部门.
@@ -14,9 +14,9 @@ import java.io.Serializable;
 public class WxCpDepart implements Serializable {
 
   private static final long serialVersionUID = -5028321625140879571L;
-  private Integer id;
+  private Long id;
   private String name;
-  private Integer parentId;
+  private Long parentId;
   private Long order;
 
   public static WxCpDepart fromJson(String json) {
